% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions_david.R
\name{AddLatentData}
\alias{AddLatentData}
\title{Add Latent Data (e.g., UMAP, MDE, PCA) to a Seurat Object}
\usage{
AddLatentData(so, latent_file, prefix, calculate_umap = F)
}
\arguments{
\item{so}{A Seurat object.}

\item{latent_file}{Character string, path to the CSV file containing the latent embeddings.}

\item{prefix}{Character string, prefix for the new \code{DimReduc} object name (e.g., "totalvi_mde").}

\item{calculate_umap}{Logical, whether to run UMAP on the newly added latent space. Default is FALSE.}
}
\value{
The modified Seurat object with the new \code{DimReduc} object.
}
\description{
This function reads latent space embeddings (e.g., from totalVI, MDE, PCA)
from a CSV file and adds them as a \code{DimReducObject} to a Seurat object.
Optionally, it can also calculate UMAP on these new embeddings.
}
