% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_functions.R
\name{ConvertS5toS3}
\alias{ConvertS5toS3}
\title{Convert Seurat V5 Object to Seurat V3 Object Structure}
\usage{
ConvertS5toS3(so, assay1 = "RNA", assay2 = "ADT")
}
\arguments{
\item{so}{A Seurat V5 object.}

\item{assay1}{Character string, name of the first assay to extract (e.g., "RNA"). Default is "RNA".}

\item{assay2}{Character string, name of the second assay to extract (e.g., "ADT"). Default is "ADT".}
}
\value{
A Seurat V3-like object.
}
\description{
This function takes a Seurat V5 object and converts it into a structure
resembling a Seurat V3 object, extracting specified assays (e.g., RNA and ADT).
}
