% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immgent_colors.R
\docType{data}
\name{immgent_colors}
\alias{immgent_colors}
\title{ImmgenT Specific Color Palettes}
\format{
A named list.
}
\usage{
immgent_colors
}
\description{
A named list of custom color palettes for ImmgenT data, organized by levels
and organ types.
}
\examples{
names(immgent_colors)
immgent_colors$level1
}
\keyword{datasets}
