% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{MakeEmbeddingRasterCache}
\alias{MakeEmbeddingRasterCache}
\title{Create a Raster Cache for Cell Embeddings}
\usage{
MakeEmbeddingRasterCache(
  so,
  reduction,
  cells = NULL,
  outfile_prefix,
  width = 2048,
  height = 2048,
  point_size = 1,
  point_alpha = 1,
  point_color = "grey"
)
}
\arguments{
\item{so}{A Seurat object.}

\item{reduction}{The name of the reduction to use (e.g., "umap").}

\item{cells}{A vector of cell IDs to include. Defaults to all cells.}

\item{outfile_prefix}{Prefix for the generated .png and .rds files.}

\item{width}{Width of the output image in pixels.}

\item{height}{Height of the output image in pixels.}

\item{point_size}{Size of the points.}

\item{point_alpha}{Transparency of the points.}

\item{point_color}{Color of the points.}
}
\value{
Invisibly returns a list containing metadata about the cache.
}
\description{
Generates a high-resolution PNG of cell embeddings (e.g., UMAP) to be used
as a fast background layer in subsequent plots.
}
