% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_functions.R
\name{MyDimPlotHighlightDensity}
\alias{MyDimPlotHighlightDensity}
\title{Generate UMAP Plot with Highlighted Cell Density}
\usage{
MyDimPlotHighlightDensity(
  seurat_object = so,
  umap_to_plot = "mde_incremental",
  group.by = sprintf("is_\%s", cl),
  split.by = NULL,
  raster = T,
  highlight_size = 0.5,
  highlight_alpha = 0.5,
  cols = rev(rainbow(10, end = 4/6))
)
}
\arguments{
\item{seurat_object}{A Seurat object. Default is \code{so}.}

\item{umap_to_plot}{Character string, name of the UMAP embedding to use. Default is "mde_incremental".}

\item{group.by}{Character string, metadata column to define the highlighted group (should be logical TRUE/FALSE). Default is \code{sprintf("is_\%s", cl)}.}

\item{split.by}{Character string, metadata column to split plots by using \code{facet_wrap}. Default is NULL.}

\item{raster}{Logical, whether to use \code{geom_scattermore} for highlighted points (TRUE) or \code{geom_point} (FALSE). Default is TRUE.}

\item{highlight_size}{Numeric, point size for highlighted cells if \code{raster} is FALSE. Default is 0.5.}

\item{highlight_alpha}{Numeric, alpha transparency for highlighted cells if \code{raster} is FALSE. Default is 0.5.}

\item{cols}{A character vector of colors to define the density color ramp. Default is \code{rev(rainbow(10, end = 4/6))}.}
}
\value{
A ggplot object.
}
\description{
This function creates a UMAP plot, highlighting a specific group of cells
by showing their density on top of a grey background of all cells.
}
