% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_functions.R
\name{MyDotPlot}
\alias{MyDotPlot}
\title{Custom DotPlot Function for Seurat Objects}
\usage{
MyDotPlot(
  object,
  features,
  assay = NULL,
  cols = c("lightgrey", "blue"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  point.shape = 21,
  alpha.range = c(0.2, 1),
  show.alpha.legend = FALSE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{features}{A vector of features to plot.}

\item{assay}{The assay to pull the data from.}

\item{cols}{A vector of colors to use.}

\item{col.min}{Minimum value for color scaling.}

\item{col.max}{Maximum value for color scaling.}

\item{dot.min}{Minimum dot size.}

\item{dot.scale}{Scale factor for dot sizes.}

\item{idents}{Identity classes to include in the plot.}

\item{group.by}{A grouping variable.}

\item{split.by}{A variable to split the plot by.}

\item{cluster.idents}{Whether to cluster identities.}

\item{scale}{Whether to scale the expression.}

\item{scale.by}{Method for scaling dot size.}

\item{scale.min}{Minimum value for dot size.}

\item{scale.max}{Maximum value for dot size.}

\item{point.shape}{Shape of the points.}

\item{alpha.range}{A vector of length 2 for alpha range.}

\item{show.alpha.legend}{Whether to show the alpha legend.}
}
\value{
A ggplot object.
}
\description{
Creates a dot plot to visualize gene expression by identity and feature. This is a modified version of Seurat's DotPlot with custom alpha and shape controls.
}
\examples{
\dontrun{
# This example demonstrates the use of ImmgenTFeaturePlots.
MyDotPlot(object = so, top_genes, assay = "RNA", group.by = "annotation_level2", dot.scale = 6, scale = T, scale.by = "radius",point.shape = 21, alpha.range = c(0, 1), show.alpha.legend = F)  +
labs(y = topic, x = NULL) + scale_fill_gradientn(colours =  colorRampPalette(c("grey",'white','red'))(10)) + theme(axis.text.x = element_text(angle = 45, hjust = 1)) + coord_flip()
} # End \dontrun
}
