% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier_functions.R
\name{MyGeneTilePlot}
\alias{MyGeneTilePlot}
\title{Create a tile plot of top gene expression for specified topics.}
\usage{
MyGeneTilePlot(
  gene_factor_matrix,
  topics = c("F68", "F27"),
  n_genes_per_topic = 15,
  alpha_range = NULL,
  title = "Dot Plot (Squares) of Gene Expression"
)
}
\arguments{
\item{gene_factor_matrix}{A matrix with genes as rows and topics as columns. Rownames must be gene names.}

\item{topics}{A character vector of the topic column names to plot.}

\item{n_genes_per_topic}{The number of top genes to select from each topic.}

\item{alpha_range}{A numeric vector of the range of values of expression to apply alpha. Default NULL is determined automatically.}

\item{title}{A character vector of the title of the plot.}
}
\value{
A ggplot object.
}
\description{
Create a tile plot of top gene expression for specified topics.
}
