% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immgent_color_palettes.R
\docType{data}
\name{mypal_list}
\alias{mypal_list}
\title{ImmgenT Specific Color Palettes}
\format{
A named list.
}
\usage{
mypal_list
}
\description{
A named list of custom color palettes for ImmgenT data, organized by levels
and organ types.
}
\examples{
names(mypal_list)
mypal_list$level1
}
\keyword{datasets}
