% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier_functions.R
\docType{data}
\name{mypal_topics}
\alias{mypal_topics}
\title{A placeholder for your color palette.
Replace this with your actual color palette.}
\format{
An object of class \code{character} of length 3.
}
\usage{
mypal_topics
}
\description{
A placeholder for your color palette.
Replace this with your actual color palette.
}
\keyword{datasets}
