% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_functions.R
\name{PlotsAfterIntegration}
\alias{PlotsAfterIntegration}
\title{Generate Plots After Seurat Integration}
\usage{
PlotsAfterIntegration(
  seurat_object = so,
  dim1 = seurat_object@reductions$umap@cell.embeddings[, 1],
  dim2 = seurat_object@reductions$umap@cell.embeddings[, 2],
  split_by = "IGT",
  genes = c("Foxp3", "Il2ra"),
  cluster_key = "ClusterSCVI_Res",
  mypal = NULL
)
}
\arguments{
\item{seurat_object}{A Seurat object. Default is \code{so}.}

\item{dim1}{Numeric vector, UMAP dimension 1 embeddings. Default is \code{so@reductions$umap@cell.embeddings[,1]}.}

\item{dim2}{Numeric vector, UMAP dimension 2 embeddings. Default is \code{so@reductions$umap@cell.embeddings[,2]}.}

\item{split_by}{Character string, metadata column to split plots by (faceting). Default is "IGT".}

\item{genes}{Character vector, genes to plot expression for. Default is \code{c("Foxp3", "Il2ra")}.}

\item{cluster_key}{Character string, key for cluster information in metadata. Default is "ClusterSCVI_Res".}

\item{mypal}{A color palette to use. Default is \code{mypal}.}
}
\value{
Prints multiple ggplot objects.
}
\description{
This function generates a series of UMAP plots often used after Seurat integration,
including plots colored by a split variable, split by a variable, gene expression, and clusters.
}
