% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions_david.R
\name{removeDuplicateColumns}
\alias{removeDuplicateColumns}
\title{Remove Duplicate Columns from a Data Frame}
\usage{
removeDuplicateColumns(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with only unique columns. Prints the names of removed duplicate columns.
}
\description{
This function identifies and removes columns from a data frame that are
exact duplicates of other columns.
}
