% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions_david.R
\name{run_limmatrend_contrasts_counfoundings}
\alias{run_limmatrend_contrasts_counfoundings}
\title{Run Limma-Trend Differential Gene Expression with Confounding Variables}
\usage{
run_limmatrend_contrasts_counfoundings(
  tmm = tmm,
  group = group,
  confoundings = confoundings,
  formula.mod.matrix = formula.mod.matrix,
  contrasts = contrasts,
  gene_symbol = gene_symbol
)
}
\arguments{
\item{tmm}{An \code{EList} object or a matrix of TMM-normalized expression data (log-transformed).}

\item{group}{A factor or vector indicating the primary grouping for differential expression.}

\item{confoundings}{A list of vectors or factors representing confounding variables.}

\item{formula.mod.matrix}{Character string, the R formula for the design matrix (e.g., "~ group + confoundings[\link{1}]").}

\item{contrasts}{A named list of character strings defining the contrasts to be made.}

\item{gene_symbol}{A vector of gene symbols corresponding to the rows of \code{tmm}.}
}
\value{
A named list of \code{topTable} results for each contrast.
}
\description{
This function performs differential gene expression analysis using \code{limma-trend}
given TMM-normalized expression data, grouping variables, confounding variables,
and specified contrasts.
}
