% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions_david.R
\name{SeSpPPVNPV}
\alias{SeSpPPVNPV}
\title{Calculate Sensitivity, Specificity, PPV, and NPV}
\usage{
SeSpPPVNPV(tp = NULL, tn = NULL, fp = NULL, fn = NULL)
}
\arguments{
\item{tp}{Numeric, true positives count.}

\item{tn}{Numeric, true negatives count.}

\item{fp}{Numeric, false positives count.}

\item{fn}{Numeric, false negatives count.}
}
\value{
A named numeric vector with se, sp, ppv, npv.
}
\description{
This function calculates sensitivity (Se), specificity (Sp), positive predictive value (PPV),
and negative predictive value (NPV) from true positive, true negative, false positive, and false negative counts.
}
\examples{
counts <- TpTnFpFn(x = c(T,T,F,F,T), y = c(T,F,T,F,T))
SeSpPPVNPV(tp = counts["tp"], tn = counts["tn"], fp = counts["fp"], fn = counts["fn"])
}
