% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_functions.R
\name{ImmgenTFeaturePlots}
\alias{ImmgenTFeaturePlots}
\title{Generate Feature Plots for ImmgenT Single-Cell Data}
\usage{
ImmgenTFeaturePlots(
  so_orig,
  so_list = Seurat::SplitObject(so_orig, split.by = "annotation_level1"),
  gene = "Cd4",
  slot = "data",
  reduction_overall = "mde2_totalvi_20241006",
  reduction_level2 = "mde_incremental",
  mypal_level1 = ZemmourLib::immgent_colors[["level1"]],
  mypal_level2 = ZemmourLib::immgent_colors[["level2"]]
)
}
\arguments{
\item{so_orig}{A Seurat object containing the overall dataset.}

\item{so_list}{A named list of Seurat objects, typically created by
\code{SplitObject(so_orig, split.by = "annotation_level1")}. Defaults to this split.}

\item{gene}{Character string, the gene to plot expression for.}

\item{slot}{Character string, the data slot to pull gene expression from (e.g., "data", "counts"). Defaults to "data".}

\item{reduction_overall}{Character string, the name of the dimensionality reduction
to use for the overall (unsplit) plots. Defaults to "mde2_totalvi_20241006".}

\item{reduction_level2}{Character string, the name of the dimensionality reduction
to use for the level 2 (split) plots. Defaults to "mde_incremental".}

\item{mypal_level1}{A character vector of colors for \code{annotation_level1}.
Defaults to \code{ZemmourLib::immgent_colors[["level1"]]}.}

\item{mypal_level2}{A character vector of colors for \code{annotation_level2}.
Defaults to \code{ZemmourLib::immgent_colors[["level2"]]}.}
}
\value{
Prints several ggplot objects and arranges them using \code{grid.arrange}.
}
\description{
This function creates a series of UMAP and Feature plots tailored for ImmgenT
single-cell data, visualizing overall expression and expression within
different annotation levels (e.g., CD4, CD8 subsets). It assumes data is
already normalized.
}
\examples{
\dontrun{
# This example demonstrates the use of ImmgenTFeaturePlots.
so_orig = NormalizeData(so_orig,assay = "RNA", normalization.method = "LogNormalize",verbose = T)
so_list = SplitObject(so_orig, split.by = "annotation_level1") # Split the seurat object
so_list[["thymocyte"]] = NULL # Remove thymocytes from the lists
mypal_level2 = immgent_colors$level2
mygene = "Cd4" # The gene to plot
pdf(sprintf("FeaturePlot_\%s.pdf",mygene), width = 20, height = 20, useDingbats = F)
ImmgenTFeaturePlots(so_orig, so_list, gene = mygene, slot = "data")
dev.off()
} # End \dontrun
}
