% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{NoGrid}
\alias{NoGrid}
\title{Theme Element: No Grid Lines}
\usage{
NoGrid(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{theme()}.}
}
\value{
A \code{theme} object for \code{ggplot2}.
}
\description{
This helper function returns a \code{ggplot2} theme element that removes both
major and minor grid lines from a plot.
}
